Huge refactor
- Refactor into multiple files
- Rewrite some components into more generic versions
  - Including some HOCs
