+++ src/misc/labels.js
@@ -1,0 +1,10 @@
+/** @typedef {'red'|'green'|'blue'|'yellow'} Label */
+
+/**
+ * List of labels that are presented to the user.
+ * @type {Label[]}
+ * @readonly 
+ */
+const LABELS = ['red', 'green', 'blue', 'yellow'];
+
+export default LABELS;
